requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.4
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cloudinary==1.21.1
coverage==5.1
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytest-django==3.9.0
python-coveralls==2.9.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
